Refactor Header component with updated Bootstrap and toggle function
This commit updates the Header component by importing the latest React Bootstrap dependencies and changing the attribute names from single quotes to double quotes. Additionally, the Navbar's responsive menu is now controlled with the "responsive-navbar-nav" id and the container class is now fluid. Finally, the handleDarkModeToggle function is slightly revised to toggle the state of darkMode.