feat: Add dark mode logo to Header
The new svg logo_dark.svg is added to the Header. When the app is in dark mode, it replaces the original logo. The app's footer background color and text color are now changed to snow and #1d1d1d respectively using css styles, which are defined in style.css of Footer component. Additionally, a new logo-long.svg is added at a different location by renaming it to logolong.svg. Finally, the package.json file is updated with a homepage URL.